Submit memos with transactions.
Memos and transactions are submitted together, at the same time. If,
as in the Espresso system, memos must be submitted after transactions
are committed, the backend should handle buffering memos and submitting
them at the appropriate time. A new `finalize` method is provided that
the backend can use to trigger memo submission after a transaction has
been committed.