Add Seahorse version to wallet persistent storage.
This isn't a complete versioning solution, but it is at least an
escape hatch. If we ever need to push an update to a released product
that changes the format of persistent storage, the updated code can
check the first field of the wallet metadata and perform a migration
if it is out of date.