Use the same viewing/freeing keys for each asset in scanner bench.
We know auditing is going to scale with the number of viewing keys
the wallet has to check. This isn't particularly interesting to
measure nor is it a common case; most wallets that audit will be
asset issuers that issue only one or a few assets.

Having the number of viewing keys scale with the number of transactions
muddled the analsysis for determining how scanning with auditing scales
with the number of transactions and the block size, so this was
removed.