Refactor loader implementations
The behavior of loading and creating keystores is a customization
point, with the KeystoreLoader trait defining the interface. This
is really nice, because it allows us to have different loaders with
different kinds of behavior. For example, we might have one loader
which only opens existing keystores, another which only creates new
ones, and another which knows how to recover a keystore from a
mnemonic phrase.

Unfortunately, over time we found ourselves with a single monolithic
loader, which was a tangle of different responsibilities. It included
an internal abstraction to handle both interactive input and preprogrammed
behavior with various combinations of a given mnemonic phrase and a
given password. Behavior was customized by setting some parameters to
None and others to Some, and non-interactive loading was awkwardly
achieved by scripting the interactive flow.

This change corrects the state of the loader by providing a handful
of non-interactive implementations, each of which accomplishes a
single task, alongside a single interactive loader that only does
interactive loading, and nothing else.

In addition, this improves the documentation and the public interface
of the loader. For example, MnemonicPasswordLogin (formerly known as
LoaderMetadata) now includes a public interface for inspecting the
metadata. There are functions to authenticate a mnemonic phrase or
password, and an interface for decrypting the mnemonic phrase given
the keystore's password.

Closes #75