Encapsulate sparse Merkle tree functionality in a type.
Since MerkleTree from jf_primitives does not support pruning the
rightmost leaf, Seahorse uses a pattern of remembering if a leaf
should be forgotten while it is the rightmost, and then forgetting
it if necessary once a new leaf has been appended.

This pattern has some subtleties, and at least two critical bugs
have been found involving not properly updating the leaf-to-forget
state. Worse, this pattern is used by both TransactionState during
the usual ledger following and in key_scan during the retroactive
catch-up scan. Recently, a bug was found in key_scan that had been
fixed a long time ago in TransactionState.

This changes removes duplication by encapsulating the leaf-forgetting
pattern in a single type which can be used by both key_scan and
TransactionState. It also makes it impossible to subvert the operations
that update the leaf-to-forget state by directly invoking an underlying
MerkleTree method, since the new data structure has an abstract interface
which automatically updates `leaf_to_forget` appropriately in all public
methods.

Finally, we add fuzz testing using quickcheck for the new abstract
data structure, which should help catch and reproduce future bugs in
this pattern at the abstract data type level, which is significantly
easier to debug than failures in situ in the ledger scanning code.