Remove Mutex protecting keystore storage.
This Mutex was already redundant with the RwLock protecting
KeystoreSharedState.

In order for this to work, I had to remove a call to `self.commit()`
inside of `AtomicKeystoreStorage::load`, so that `load` could take
`&self` instead of `&mut self`. This is as it should be. For
consistency, I also removed a `commit` in `create`, and pushed the
responsibility for calling `commit` up to the caller. This makes
sense: `AtomicKeystoreStorage` is simulating the same kind of
interface provided by the primitive atomic-store log types, where
`commit` is always called by the owner of the log, not the log
itself, right before calling `commit_version` on the overall atomic
store. This makes it much easier to synchronize commits on the individual
log and on the overal store, given that the store is now separated from
the logs.