Remove `block_on` in destructor.
As [I recently learned](https://github.com/EspressoSystems/seahorse/issues/285),
using `block_on` in a destructor is full of tricks and subtleties.
While this particular use was OK since the future could never pend,
it was a recipe for problems in the future. Since the future (meta_store.revert())
could never pend, it is better to prove that in the type system by
converting `revert` to a synchronous function, which incidentally
makes it more consistent with the `revert` functions on other models.

At the same time, I also changed `commit` and `revert` for meta store
to return results instead of unwrapping, which is also more consistent
with the other `commit` and `revert` functions.